re-generated docs, updated CONTRIBUTING.md to reflect new format
