+++ docs/content/docs/tools/image_generator.mdx
@@ -37,21 +37,6 @@ Create high-quality images using DALL-E. Configure resolution, quality, style, a
 
 ### `dalle_generate`
 
-Generate images using OpenAI
-
-#### Input Parameters
-
-| Parameter | Type | Required | Description |
-| --------- | ---- | -------- | ----------- |
-| `prompt` | string | Yes | A text description of the desired image(s) |
-
-#### Output
-
-| Parameter | Type |
-| --------- | ---- |
-| `content` | string |
-| `image` | string |
-| `metadata` | string |
