+++ docs/content/docs/tools/jina.mdx
@@ -52,12 +52,15 @@ Transform web content into clean, readable text using Jina AI
 
 ## Tools
 
-### `jina_readurl`
+### `jina_read_url`
+
+Extract and process web content into clean, LLM-friendly text using Jina AI Reader. Supports advanced content parsing, link gathering, and multiple output formats with configurable processing options.
 
 #### Input Parameters
 
 | Parameter | Type | Required | Description |
 | --------- | ---- | -------- | ----------- |
+| `url` | string | Yes | The URL to read and convert to markdown |
 
 #### Output
