+++ docs/content/docs/tools/notion.mdx
@@ -69,6 +69,7 @@ Create a new page in Notion
 
 | Parameter | Type | Required | Description |
 | --------- | ---- | -------- | ----------- |
+| `parentType` | string | Yes | Type of parent:  |
 
 #### Output
