+++ docs/content/docs/tools/thinking.mdx
@@ -32,10 +32,13 @@ Adds a step where the model explicitly outlines its thought process before proce
 
 ### `thinking_tool`
 
+Processes a provided thought/instruction, making it available for subsequent steps.
+
 #### Input Parameters
 
 | Parameter | Type | Required | Description |
 | --------- | ---- | -------- | ----------- |
+| `thought` | string | Yes | The thought process or instruction provided by the user in the Thinking Step block. |
 
 #### Output
