+++ docs/content/docs/tools/typeform.mdx
@@ -44,6 +44,14 @@ Retrieve form responses from Typeform
 
 #### Output
 
+| Parameter | Type |
+| --------- | ---- |
+| `total_items` | string |
+| `answers` | string |
+| `type` | string |
+| `hidden` | string |
+| `calculated` | string |
+| `variables` | string |
 
 ### `typeform_files`
