+++ docs/content/docs/tools/vision.mdx
@@ -40,13 +40,21 @@ Process visual content with customizable prompts to extract insights and informa
 
 ### `vision_tool`
 
+Process and analyze images using advanced vision models. Capable of understanding image content, extracting text, identifying objects, and providing detailed visual descriptions.
+
 #### Input Parameters
 
 | Parameter | Type | Required | Description |
 | --------- | ---- | -------- | ----------- |
+| `apiKey` | string | Yes | API key for the selected model provider |
 
 #### Output
 
+| Parameter | Type |
+| --------- | ---- |
+| `content` | string |
+| `model` | string |
+| `tokens` | string |
