+++ docs/content/docs/tools/whatsapp.mdx
@@ -37,6 +37,7 @@ Send WhatsApp messages
 
 | Parameter | Type | Required | Description |
 | --------- | ---- | -------- | ----------- |
+| `phoneNumber` | string | Yes | Recipient phone number with country code |
 
 #### Output
