+++ sim/tools/jina/index.ts
@@ -1,3 +1,3 @@
-import { readUrlTool } from './reader'
+import { readUrlTool } from './read_url'
 
 export { readUrlTool }
