+++ sim/tools/notion/index.ts
@@ -1,5 +1,6 @@
 import { notionCreatePageTool } from './create_page'
 import { notionReadTool } from './read'
 import { notionWriteTool } from './write'
+import { notionUpdatePageTool } from './update_page'
 
-export { notionReadTool, notionWriteTool, notionCreatePageTool }
+export { notionReadTool, notionWriteTool, notionCreatePageTool, notionUpdatePageTool }
