+++ sim/tools/pinecone/generate_embeddings.ts
