+++ sim/tools/registry.ts
@@ -53,10 +53,10 @@ export const tools: Record<string, ToolConfig> = {
   file_parser: fileParseTool,
   firecrawl_scrape: scrapeTool,
   google_search: googleSearchTool,
-  jina_readurl: readUrlTool,
+  jina_read_url: readUrlTool,
   linkup_search: linkupSearchTool,
   slack_message: slackMessageTool,
-  github_repoinfo: githubRepoInfoTool,
+  github_repo_info: githubRepoInfoTool,
   github_latest_commit: githubLatestCommitTool,
   serper_search: serperSearch,
   tavily_search: tavilySearchTool,
