+++ sim/tools/thinking/tool.ts
