+++ sim/tools/vision/tool.ts
